+++ client/src/main.jsx
@@ -1,10 +1,13 @@
 import { createRoot } from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 import App from "./App.jsx";
+import AppErrorBoundary from "./components/AppErrorBoundary.jsx";
 import "./style.css";
 
 createRoot(document.getElementById("root")).render(
-  <BrowserRouter>
-    <App />
-  </BrowserRouter>
+  <AppErrorBoundary>
+    <BrowserRouter>
+      <App />
+    </BrowserRouter>
+  </AppErrorBoundary>
 );